+++ src/domain/ports/IShortUrlsRepository.ts
@@ -3,4 +3,5 @@ import type { IShortUrl } from "../entities/ShortUrl";
 export interface IShortUrlsRepository {
   create(id: string, originalUrl: string): Promise<IShortUrl>;
   findById(id: string): Promise<IShortUrl | null>;
+  updateClickCount(id: string): Promise<null>;
 }