+++ src/domain/ports/IUsersRepository.ts
@@ -3,4 +3,5 @@ import type { IUser } from "../entities/User";
 export interface IUsersRepository {
   create(username: string, password: string): Promise<IUser>;
   findByUsername(username: string): Promise<IUser | null>;
+  findById(id: string): Promise<IUser | null>;
 }