+++ src/infrastructure/database/typeorm/migrations/1744260818169-createUsersTable.ts
@@ -1,6 +1,6 @@
 import { MigrationInterface, QueryRunner, Table, TableIndex } from "typeorm";
 
-export class CreateUsersTable1744261632839 implements MigrationInterface {
+export class CreateUsersTable1744260818169 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.createTable(
       new Table({