+++ src/infrastructure/database/typeorm/migrations/1744261632839-createShortUrlsTable.ts
@@ -1,6 +1,6 @@
 import { MigrationInterface, QueryRunner, Table, TableIndex } from "typeorm";
 
-export class CreateShortUrlsTable1744260818169 implements MigrationInterface {
+export class CreateShortUrlsTable1744261632839 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.createTable(
       new Table({
@@ -19,6 +19,16 @@ export class CreateShortUrlsTable1744260818169 implements MigrationInterface {
             type: "varchar",
           },
           {
+            name: "click_count",
+            type: "int",
+            default: 0,
+          },
+          {
+            name: "user_id",
+            type: "uuid",
+            isNullable: true,
+          },
+          {
             name: "created_at",
             type: "timestamp",
             default: "now()",
@@ -32,6 +42,15 @@ export class CreateShortUrlsTable1744260818169 implements MigrationInterface {
             name: "deleted_at",
             type: "timestamp",
             isNullable: true,
+          },
+        ],
+        foreignKeys: [
+          {
+            name: "FK_short_urls_user_id",
+            referencedTableName: "users",
+            referencedColumnNames: ["id"],
+            columnNames: ["user_id"],
+            onDelete: "SET NULL",
           },
         ],
       }),