+++ src/infrastructure/http/controllers/redirect-url-controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import { AppError } from "../../../domain/errors/AppError";
-import { GetShortUrlById } from "../../../aplication/get-short-url-by-id";
+import { GetShortUrlById } from "../../../application/get-short-url-by-id";
 
 export class RedirectUrlController {
   async handle(request: Request, response: Response) {
@@ -19,6 +19,6 @@ export class RedirectUrlController {
       throw new AppError("URL not found", "SHORT_URL_NOT_FOUND", 404);
     }
 
-    response.status(302).redirect(shortUrl.originalUrl);
+    response.status(302).redirect(shortUrl.original_url);
   }
 }