+++ src/infrastructure/http/routes/short-urls.routes.ts
@@ -1,9 +1,14 @@
 import { Router } from "express";
 
 import { CreateShortUrlController } from "../controllers/create-short-url-controller";
+import { ensureOptionalAuthenticated } from "../middlewares/ensure-optional-authenticated";
 
 export const shortUrlsRoutes = Router();
 
 const createShortUrlController = new CreateShortUrlController();
 
-shortUrlsRoutes.post("/", createShortUrlController.handle);
+shortUrlsRoutes.post(
+  "/",
+  ensureOptionalAuthenticated,
+  createShortUrlController.handle,
+);